pyproject.toml
[project]
name = "BayesianOptimization"
version = "0.1.0"
description = "Bayesian optimization core library"
requires-python = ">=3.12"

dependencies = [
	"jax",
	"jaxlib",
	"optax",
	"numpy",
	"absl-py",
	"pytest"
]
